Properly fix cFile Warnings

+++ src/OSSupport/File.cpp
@@ -180,7 +180,7 @@ int cFile::Write(const void * iBuffer, size_t iNumBytes)
 
 
 
-int cFile::Seek (int iPosition)
+long cFile::Seek (int iPosition)
 {
 	ASSERT(IsOpen());
 	
@@ -193,7 +193,7 @@ int cFile::Seek (int iPosition)
 	{
 		return -1;
 	}
-	return (int)ftell(m_File);
+	return ftell(m_File);
 }
 
 
@@ -201,7 +201,7 @@ int cFile::Seek (int iPosition)
 
 
 
-ssize_t cFile::Tell (void) const
+long cFile::Tell (void) const
 {
 	ASSERT(IsOpen());
 	
@@ -217,7 +217,7 @@ ssize_t cFile::Tell (void) const
 
 
 
-ssize_t cFile::GetSize(void) const
+long cFile::GetSize(void) const
 {
 	ASSERT(IsOpen());
 	
@@ -226,7 +226,7 @@ ssize_t cFile::GetSize(void) const
 		return -1;
 	}
 	
-	ssize_t CurPos = Tell();
+	long CurPos = Tell();
 	if (CurPos < 0)
 	{
 		return -1;
@@ -235,7 +235,7 @@ ssize_t cFile::GetSize(void) const
 	{
 		return -1;
 	}
-	ssize_t res = Tell();
+	long res = Tell();
 	if (fseek(m_File, (long)CurPos, SEEK_SET) != 0)
 	{
 		return -1;
@@ -361,7 +361,7 @@ bool cFile::IsFile(const AString & a_Path)
 
 
 
-ssize_t cFile::GetSize(const AString & a_FileName)
+long cFile::GetSize(const AString & a_FileName)
 {
 	struct stat st;
 	if (stat(a_FileName.c_str(), &st) == 0)
